+++ src/hooks/index.ts
@@ -1,3 +1,8 @@
+export * from './useAnnotationEvents';
+export * from './useBigCalendarEvents';
 export * from './useCalendarEvents';
 export * from './useCalendarRange';
+export * from './useColors';
+export * from './useEventFrames';
+export * from './useIntervalSelection';
 export * from './useLocalizer';
