+++ src/hooks/useIntervalSelection.ts
@@ -1,52 +1,8 @@
 import dayjs from 'dayjs';
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 
 import { Range } from '../types';
-
-/**
- * Key Press
- */
-export const useKeyPress = (key: string, onKeyPressed: (pressed: boolean) => void) => {
-  /**
-   * Key down
-   */
-  const keydownListener = useCallback(
-    (e: KeyboardEvent) => {
-      if (e.key === 'Shift') {
-        onKeyPressed(true);
-      }
-    },
-    [onKeyPressed]
-  );
-
-  /**
-   * Key up
-   */
-  const keyupListener = useCallback(
-    (e: KeyboardEvent) => {
-      if (e.key === 'Shift') {
-        onKeyPressed(false);
-      }
-    },
-    [onKeyPressed]
-  );
-
-  /**
-   * Event for Key down
-   */
-  useEffect(() => {
-    window.addEventListener('keydown', keydownListener, true);
-    return () => window.removeEventListener('keydown', keydownListener, true);
-  }, [keydownListener]);
-
-  /**
-   * Event for Key up
-   */
-  useEffect(() => {
-    window.addEventListener('keyup', keyupListener, true);
-    return () => window.removeEventListener('keyup', keyupListener, true);
-  }, [keyupListener]);
-};
+import { useKeyPress } from './useKeyPress';
 
 /**
  * Interval Selection
